requirements.txt
adal==1.2.7
aiokafka==0.7.2
alembic==1.8.1
anyio==3.6.1
argcomplete==2.0.0
attrs==22.1.0
azure-common==1.1.28
azure-core==1.24.2
azure-graphrbac==0.61.1
azure-mgmt-authorization==2.0.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==10.0.0
azure-mgmt-resource==21.1.0
azure-mgmt-storage==20.0.0
azureml-core==1.43.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.2
boto3==1.24.44
botocore==1.27.44
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
contextlib2==21.6.0
cryptography==37.0.4
cycler==0.11.0
databricks-cli==0.17.0
distlib==0.3.5
docker==5.0.3
entrypoints==0.4
fastapi==0.79.0
filelock==3.7.1
flake8==4.0.1
Flask==2.2.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-auth==2.9.1
google-cloud-core==2.3.2
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.47.0
gunicorn==20.1.0
h11==0.13.0
humanfriendly==10.0
hyperopt==0.2.7
idna==3.3
importlib-metadata==4.2.0
importlib-resources==5.9.0
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
jsonpickle==2.2.0
kafka-python==2.0.2
kiwisolver==1.4.4
knack==0.9.0
kubernetes==24.2.0
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib==3.3.4
mccabe==0.6.1
mlflow==1.27.0
mlserver==1.1.0
mlserver-mlflow==1.1.0
msal==1.18.0
msal-extensions==1.0.0
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
paramiko==2.11.0
pathspec==0.9.0
Pillow==9.2.0
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
portalocker==2.5.1
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
protobuf==4.21.4
py==1.11.0
py-grpc-prometheus==0.7.0
py4j==0.10.9.5
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.4.0
PyNaCl==1.5.0
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.9
pysftp==0.2.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
querystring-parser==1.2.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.5.4
seaborn==0.11.2
SecretStorage==3.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
starlette==0.20.4
starlette-exporter==0.13.0
tabulate==0.8.10
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.0
typing-extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
virtualenv==20.16.2
websocket-client==1.3.3
Werkzeug==2.2.1
zipp==3.8.1
